@@ -24,17 +24,21 @@
         "@babel/eslint-parser": "^7.12.16",
         "@babel/preset-env": "^7.23.2",
         "@types/uuid": "^9.0.6",
+        "@typescript-eslint/eslint-plugin": "^6.12.0",
+        "@typescript-eslint/parser": "^6.12.0",
         "@vue/cli-plugin-babel": "~5.0.0",
         "@vue/cli-plugin-eslint": "~5.0.0",
         "@vue/cli-plugin-router": "~5.0.0",
         "@vue/cli-plugin-vuex": "~5.0.0",
         "@vue/cli-service": "^5.0.8",
         "@vue/compiler-sfc": "^3.3.7",
-        "eslint": "^8.51.0",
+        "eslint": "^8.54.0",
         "eslint-config-prettier": "^8.3.0",
         "eslint-plugin-prettier": "^5.0.1",
         "eslint-plugin-vue": "^9.17.0",
-        "prettier": "^3.0.3"
+        "prettier": "^3.0.3",
+        "typescript": "^5.3.2",
+        "vite": "^5.0.2"
     },
     "eslintConfig": {
         "root": true,
