@@ -3,11 +3,12 @@ module.exports = {
     env: {
         node: true,
         browser: true,
+        es2021: true,
     },
     extends: [
-        'plugin:vue/vue3-recommended',
         'eslint:recommended',
-        'plugin:prettier/recommended',
+        'plugin:@typescript-eslint/recommended',
+        'prettier',
     ],
     rules: {
         'prettier/prettier': [
@@ -23,8 +24,10 @@ module.exports = {
     globals: {
         $nuxt: true,
     },
+    parser: '@typescript-eslint/parser',
     parserOptions: {
-        parser: '@babel/eslint-parser',
-        requireConfigFile: false,
+        ecmaVersion: 'latest',
+        sourceType: 'module',
     },
+    plugins: ['@typescript-eslint', 'prettier'],
 };
