@@ -1,1 +1,1 @@
-export const currenciesMap = { 840: 'USD', 978: 'EUR', 980: 'UAH' }
+export const currenciesMap = { 840: 'USD', 978: 'EUR', 980: 'UAH' };
