@@ -14,7 +14,7 @@
 <script>
 export default {
     name: 'FooterComponent',
-}
+};
 </script>
 
 <style scoped>
