@@ -1,6 +1,6 @@
-import CurrencyInput from '@/components/UI/CurrencyInput.vue'
-import ButtonForConvert from '@/components/UI/Button.vue'
-import NavButton from '@/components/UI/NavButton'
-import SmallButton from '@/components/UI/SmallButton'
+import CurrencyInput from '@/components/UI/CurrencyInput.vue';
+import ButtonForConvert from '@/components/UI/Button.vue';
+import NavButton from '@/components/UI/NavButton';
+import SmallButton from '@/components/UI/SmallButton';
 
-export default [CurrencyInput, ButtonForConvert, NavButton, SmallButton]
+export default [CurrencyInput, ButtonForConvert, NavButton, SmallButton];
