@@ -1,3 +1,3 @@
-import VFocus from './VFocus'
+import VFocus from './VFocus';
 
-export default [VFocus]
+export default [VFocus];
