@@ -7,8 +7,8 @@
 </template>
 
 <script setup>
-import Navbar from '@/components/NavBar.vue'
-import FooterComponent from '@/components/FooterComponent.vue'
+import Navbar from '@/components/NavBar.vue';
+import FooterComponent from '@/components/FooterComponent.vue';
 </script>
 
 <style scoped>
