@@ -5,13 +5,13 @@
 </template>
 
 <script>
-import InputForm from '@/components/InputForm.vue'
+import InputForm from '@/components/InputForm.vue';
 export default {
     name: 'ConverterPage',
     components: {
         InputForm,
     },
-}
+};
 </script>
 
 <style scoped>
