@@ -12,7 +12,7 @@
 <script>
 export default {
     name: 'IntroPage',
-}
+};
 </script>
 
 <style scoped>
