@@ -1,8 +1,8 @@
-import { createStore } from 'vuex'
-import amountModule from './convertModule'
+import { createStore } from 'vuex';
+import amountModule from './convertModule';
 
 export default createStore({
     modules: {
         convert: amountModule,
     },
-})
+});
